+++ src/main/java/controller/UserController.java
@@ -4,7 +4,7 @@ import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.HttpRequestUtils;
-import webserver.HttpRequest;
+import webserver.http.HttpRequest;
 
 import java.util.Map;
