+++ src/main/java/webserver/RequestHandler.java
@@ -8,6 +8,7 @@ import controller.UserController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.HttpRequestUtils;
+import webserver.http.HttpRequest;
 
 
 public class RequestHandler implements Runnable {
